+++ tests/test_timing.py
@@ -3,7 +3,7 @@ import pytest
 import scipy.ndimage
 import torch
 
-from whisper.timing import dtw_cpu, dtw_cuda, median_filter
+from whisperng.timing import dtw_cpu, dtw_cuda, median_filter
 
 sizes = [
     (10, 20),